+++ common/src/main/java/earth/terrarium/argonauts/common/network/messages/ServerboundChatWindowPacket.java
@@ -73,7 +73,7 @@ public record ServerboundChatWindowPacket(String message,
         }
     }
 
-    private static void sendMessage(ServerPlayer player, Group<?, ?> group, String message, MessageChannel channel) {
+    public static void sendMessage(ServerPlayer player, Group<?, ?> group, String message, MessageChannel channel) {
         ChatMessage chatMessage = new ChatMessage(
             player.getGameProfile(),
             message,